enh: increased tag size limit to 31
Started using UInt32 instead of UInt16 as this does not seem to affect the speed of the program.

+++ src/barcodes.rs
@@ -14,12 +14,12 @@ type AlignedTags<'a> = Vec<(usize, &'a Tag)>;
 
 /// Look for a tag in a DNA/cDNA read using the bitap algorithm
 ///
-/// NOTE: Currently tag lenght must be 16 bp or lower
+/// NOTE: Currently tag lenght must be 32 bp or lower
 fn match_pattern(sequence: &[u8], tag: &Tag) -> Option<usize> {
     // Setup initial arrays (1 for exact match, 1 for each allowed mismatch).
     // NOTE: this allocation seems to be trivial speed-wise.
-    let mut state: [u16; MAX_TAG_MISM + 1] = [!1u16; MAX_TAG_MISM + 1];
-    let match_mask: u16 = 1u16 << tag.len;
+    let mut state: [u32; MAX_TAG_MISM + 1] = [!1u32; MAX_TAG_MISM + 1];
+    let match_mask: u32 = 1u32 << tag.len;
 
     for (i, base) in sequence.iter().enumerate() {
         let mut old_state = state[0];
